fix(core): fix out of bounds access on stream restart
When the capturer fails with a permanent error from the OS, we
are restarting it and during the restart we were accessing the first
available source without checking if there are no sources. This is
wrong because if the captuer is in an error state it could return
0 available sources.

Stream::start_capture has been modified now to return an error.