fix: fix out of bounds access on stream start_capture (#95)
* improvement(frontend): add loading spinner when selecting content

* chore(backend): update mock data

* fix(core): fix out of bounds access on stream restart

When the capturer fails with a permanent error from the OS, we
are restarting it and during the restart we were accessing the first
available source without checking if there are no sources. This is
wrong because if the captuer is in an error state it could return
0 available sources.

Stream::start_capture has been modified now to return an error.

* improvement: show better error messages when core has stopped

* chore: make form more engaging

* fix: add spiner state var

* chore: minor styling and mock data fixes

* chore: clean unused code

* review: restart core instead of requesting to restart

---------

Co-authored-by: konsalex <konstantinos.alexoglou@grafana.com>