fix: missing logs after restart (#99)
* improvement: remove spammy ping log

* fix: remove sensitive data from logs

* fix: empty logs after restart

Fixes an issue where after restarting the app the logs where empty.
This was happening because the env variable RUST_LOG was set from the
previous run and as a result we were using WARN as the level.

* chore: add max size for logs

* improvement: enhance logs attached to sentry

The sentry entry has been extended with system, user and app data.
Also we are logging when a call has started for easier filtering during
debugging.

* chore: fix clippy