feat: allow bigger camera windows
Changes the camera sharing to allow bigger resolution, in particular
now we do 720, 360 and 216.

In the camera window we now allow setting three sizes and in each
size we use one of the layers. This will allow users to have a call
with just the camera if they don't want to screen share.